requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
openai==1.3.7
requests==2.31.0
pydantic>=2.11.2
typing-extensions>=4.6.0
SpeechRecognition==3.10.0 
pyttsx3==2.90 
pyaudio==0.2.11


# New voice processing packages
openai-whisper==20231117
TTS==0.22.0
torch>=1.9.0
torchaudio>=0.9.0
soundfile>=0.12.1
librosa>=0.9.0